fixed a bug

+++ auth.ts
@@ -94,26 +94,28 @@ export const config = {
         }
 
         if (trigger === 'signIn' || trigger === 'signUp') {
+          // Get the cookies
           const cookiesObject = await cookies()
+
+          // Get the not logged in cart from the cookies
           const sessionCartId = cookiesObject.get('sessionCartId')?.value
 
-          if (sessionCartId) {
-            const sessionCart = await prisma.cart.findFirst({
-              where: { sessionCartId },
+          // If the not logged in user has added anything to its cart , it would create a record in database and if he's added nothing then it wouldn't create
+          const sessionCart = await prisma.cart.findFirst({
+            where: { sessionCartId },
+          })
+
+          if (sessionCartId && sessionCart) {
+            // Delete current user cart
+            await prisma.cart.deleteMany({
+              where: { userId: user.id },
             })
 
-            if (sessionCart) {
-              // Delete current user cart
-              await prisma.cart.deleteMany({
-                where: { userId: user.id },
-              })
-
-              // Assign new cart
-              await prisma.cart.update({
-                where: { id: sessionCart.id },
-                data: { userId: user.id },
-              })
-            }
+            // Assign new cart
+            await prisma.cart.update({
+              where: { id: sessionCart.id },
+              data: { userId: user.id },
+            })
           }
         }
       }
